Move text tests to separate file
