+++ tests/test_text.py
@@ -1,0 +1,74 @@
+from datetime import timedelta
+from typing import Optional
+
+import hypothesis.strategies as st
+import numpy as np
+import pandas as pd
+import pandas.testing as tm
+import pyarrow as pa
+import pytest
+from hypothesis import given, settings
+
+import fletcher as fr
+
+
+@settings(deadline=timedelta(milliseconds=1000))
+@given(data=st.lists(st.one_of(st.text(), st.none())))
+def test_text_cat(data, fletcher_variant, fletcher_variant_2):
+    if any("\x00" in x for x in data if x):
+        # pytest.skip("pandas cannot handle \\x00 characters in tests")
+        # Skip is not working properly with hypothesis
+        return
+    ser_pd = pd.Series(data, dtype=str)
+    arrow_data = pa.array(data, type=pa.string())
+    if fletcher_variant == "chunked":
+        fr_array = fr.FletcherChunkedArray(arrow_data)
+    else:
+        fr_array = fr.FletcherContinuousArray(arrow_data)
+    ser_fr = pd.Series(fr_array)
+    if fletcher_variant_2 == "chunked":
+        fr_other_array = fr.FletcherChunkedArray(arrow_data)
+    else:
+        fr_other_array = fr.FletcherContinuousArray(arrow_data)
+    ser_fr_other = pd.Series(fr_other_array)
+
+    result_pd = ser_pd.str.cat(ser_pd)
+    result_fr = ser_fr.fr_text.cat(ser_fr_other)
+    result_fr = result_fr.astype(object)
+    # Pandas returns np.nan for NA values in cat, keep this in line
+    result_fr[result_fr.isna()] = np.nan
+    tm.assert_series_equal(result_fr, result_pd)
+
+
+def _optional_len(x: Optional[str]) -> int:
+    if x is not None:
+        return len(x)
+    else:
+        return 0
+
+
+@settings(deadline=timedelta(milliseconds=1000))
+@given(data=st.lists(st.one_of(st.text(), st.none())))
+@pytest.mark.xfail(reason="Not implemented")
+def test_text_zfill(data, fletcher_variant):
+    if any("\x00" in x for x in data if x):
+        # pytest.skip("pandas cannot handle \\x00 characters in tests")
+        # Skip is not working properly with hypothesis
+        return
+    ser_pd = pd.Series(data, dtype=str)
+    max_str_len = ser_pd.map(_optional_len).max()
+    if pd.isna(max_str_len):
+        max_str_len = 0
+    arrow_data = pa.array(data, type=pa.string())
+    if fletcher_variant == "chunked":
+        fr_array = fr.FletcherChunkedArray(arrow_data)
+    else:
+        fr_array = fr.FletcherContinuousArray(arrow_data)
+    ser_fr = pd.Series(fr_array)
+
+    result_pd = ser_pd.str.zfill(max_str_len + 1)
+    result_fr = ser_fr.fr_text.zfill(max_str_len + 1)
+    result_fr = result_fr.astype(object)
+    # Pandas returns np.nan for NA values in cat, keep this in line
+    result_fr[result_fr.isna()] = np.nan
+    tm.assert_series_equal(result_fr, result_pd)
